backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
Pillow==10.1.0
numpy==1.24.3
opencv-python==4.8.1.78
rembg==2.0.50
torch==2.1.0
transformers==4.35.0
python-multipart==0.0.6
scipy==1.11.4
